I heard you like refactoring, so I put refactoring in your refactoring
